requirements.txt
asgiref==3.3.4
attrs==21.4.0
certifi==2021.10.8
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
djangorestframework==3.14.0
drf-yasg==1.20.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.5.1
MarkupSafe==2.1.1
mysqlclient==2.0.3
packaging==21.3
pyparsing==3.0.8
pyrsistent==0.18.1
pytz==2021.1
PyYAML==5.3.1
requests==2.25.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.1
swagger-spec-validator==2.7.4
uritemplate==4.1.1
urllib3==1.26.9
